Move handler trace logic into handlertrace package
