+++ lambda/handler.go
@@ -7,6 +7,8 @@ import (
 	"encoding/json"
 	"fmt"
 	"reflect"
+
+	"github.com/aws/aws-lambda-go/lambda/handlertrace"
 )
 
 type Handler interface {
@@ -70,42 +72,6 @@ func validateReturns(handler reflect.Type) error {
 	return nil
 }
 
-// HandlerTrace allows handlers which wrap the return value of NewHandler to
-// access to the request and response events.
-type HandlerTrace struct {
-	RequestEvent  func(context.Context, interface{})
-	ResponseEvent func(context.Context, interface{})
-}
-
-func callbackCompose(f1, f2 func(context.Context, interface{})) func(context.Context, interface{}) {
-	return func(ctx context.Context, event interface{}) {
-		if nil != f1 {
-			f1(ctx, event)
-		}
-		if nil != f2 {
-			f2(ctx, event)
-		}
-	}
-}
-
-type handlerTraceKey struct{}
-
-// WithHandlerTrace adds callbacks to the provided context which allows handlers
-// which wrap the return value of NewHandler to access to the request and
-// response events.
-func WithHandlerTrace(ctx context.Context, trace HandlerTrace) context.Context {
-	existing := contextHandlerTrace(ctx)
-	return context.WithValue(ctx, handlerTraceKey{}, HandlerTrace{
-		RequestEvent:  callbackCompose(existing.RequestEvent, trace.RequestEvent),
-		ResponseEvent: callbackCompose(existing.ResponseEvent, trace.ResponseEvent),
-	})
-}
-
-func contextHandlerTrace(ctx context.Context) HandlerTrace {
-	trace, _ := ctx.Value(handlerTraceKey{}).(HandlerTrace)
-	return trace
-}
-
 // NewHandler creates a base lambda handler from the given handler function. The
 // returned Handler performs JSON deserialization and deserialization, and
 // delegates to the input handler function.  The handler function parameter must
@@ -132,7 +98,7 @@ func NewHandler(handlerFunc interface{}) Handler {
 
 	return lambdaHandler(func(ctx context.Context, payload []byte) (interface{}, error) {
 
-		trace := contextHandlerTrace(ctx)
+		trace := handlertrace.ContextHandlerTrace(ctx)
 
 		// construct arguments
 		var args []reflect.Value
